Move to recipient service and add tests
