+++ src/routes/recipient/handlers.test.js
@@ -6,10 +6,14 @@ import {
   getGoalsByRecipient,
   getGoalsByIdandRecipient,
   getRecipientAndGrantsByUser,
+  getRecipientLeadership,
+  getRecipientLeadershipHistory,
 } from './handlers';
 import {
   getGoalsByActivityRecipient,
   recipientById,
+  recipientLeadership,
+  recipientLeadershipHistory,
   recipientsByName,
   recipientsByUserId,
 } from '../../services/recipient';
@@ -28,6 +32,8 @@ jest.mock('../../services/recipient', () => ({
   getUserReadRegions: jest.fn(),
   updateRecipientGoalStatusById: jest.fn(),
   recipientsByUserId: jest.fn(),
+  recipientLeadership: jest.fn(),
+  recipientLeadershipHistory: jest.fn(),
 }));
 
 jest.mock('../../services/goals', () => ({
@@ -251,6 +257,156 @@ describe('getGoalsByActivityRecipient', () => {
     };
     getUserReadRegions.mockResolvedValue([2]);
     await getGoalsByRecipient(req, mockResponse);
+    expect(mockResponse.sendStatus).toHaveBeenCalledWith(403);
+  });
+});
+
+describe('getRecipientLeadership', () => {
+  const mockResponse = {
+    attachment: jest.fn(),
+    json: jest.fn(),
+    send: jest.fn(),
+    sendStatus: jest.fn(),
+    status: jest.fn(() => ({
+      end: jest.fn(),
+    })),
+  };
+  it('retrieves goals by recipient', async () => {
+    const req = {
+      params: {
+        recipientId: 100000,
+        regionId: 1,
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue({});
+    getUserReadRegions.mockResolvedValue([1]);
+    recipientLeadership.mockResolvedValue([]);
+    await getRecipientLeadership(req, mockResponse);
+    expect(mockResponse.json).toHaveBeenCalledWith([]);
+  });
+
+  it('returns a 404 when a recipient can\'t be found', async () => {
+    const req = {
+      params: {
+        recipientId: 14565,
+        regionId: 1,
+      },
+      query: {
+        'region.in': 1,
+        modelType: 'grant',
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue(null);
+    getUserReadRegions.mockResolvedValue([1]);
+    recipientLeadership.mockResolvedValue(null);
+    await getRecipientLeadership(req, mockResponse);
+    expect(mockResponse.sendStatus).toHaveBeenCalledWith(NOT_FOUND);
+  });
+
+  it('returns a 500 on error', async () => {
+    const req = {
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue({});
+    await getRecipientLeadership(req, mockResponse);
+    expect(mockResponse.status).toHaveBeenCalledWith(INTERNAL_SERVER_ERROR);
+  });
+
+  it('returns a 403 on region permissions', async () => {
+    const req = {
+      params: {
+        recipientId: 14565,
+        regionId: 1,
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    getUserReadRegions.mockResolvedValue([2]);
+    await getRecipientLeadership(req, mockResponse);
+    expect(mockResponse.sendStatus).toHaveBeenCalledWith(403);
+  });
+});
+
+describe('getRecipientLeadershipHistory', () => {
+  const mockResponse = {
+    attachment: jest.fn(),
+    json: jest.fn(),
+    send: jest.fn(),
+    sendStatus: jest.fn(),
+    status: jest.fn(() => ({
+      end: jest.fn(),
+    })),
+  };
+  it('retrieves goals by recipient', async () => {
+    const req = {
+      params: {
+        recipientId: 100000,
+        regionId: 1,
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue({});
+    getUserReadRegions.mockResolvedValue([1]);
+    recipientLeadershipHistory.mockResolvedValue([]);
+    await getRecipientLeadershipHistory(req, mockResponse);
+    expect(mockResponse.json).toHaveBeenCalledWith([]);
+  });
+
+  it('returns a 404 when a recipient can\'t be found', async () => {
+    const req = {
+      params: {
+        recipientId: 14565,
+        regionId: 1,
+      },
+      query: {
+        'region.in': 1,
+        modelType: 'grant',
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue(null);
+    getUserReadRegions.mockResolvedValue([1]);
+    recipientLeadershipHistory.mockResolvedValue(null);
+    await getRecipientLeadershipHistory(req, mockResponse);
+    expect(mockResponse.sendStatus).toHaveBeenCalledWith(NOT_FOUND);
+  });
+
+  it('returns a 500 on error', async () => {
+    const req = {
+      session: {
+        userId: 1000,
+      },
+    };
+    recipientById.mockResolvedValue({});
+    await getRecipientLeadershipHistory(req, mockResponse);
+    expect(mockResponse.status).toHaveBeenCalledWith(INTERNAL_SERVER_ERROR);
+  });
+
+  it('returns a 403 on region permissions', async () => {
+    const req = {
+      params: {
+        recipientId: 14565,
+        regionId: 1,
+      },
+      session: {
+        userId: 1000,
+      },
+    };
+    getUserReadRegions.mockResolvedValue([2]);
+    await getRecipientLeadershipHistory(req, mockResponse);
     expect(mockResponse.sendStatus).toHaveBeenCalledWith(403);
   });
 });
